fixed the snapping bug

+++ src/gui/gui_main.hpp
@@ -12,4 +12,4 @@ std::vector<sf::RectangleShape> init_squares();
 std::unordered_map<int, std::shared_ptr<sf::Texture>> init_textures(); 
 PieceSprite* get_piece_at_position(const sf::Vector2i&, std::vector<PieceSprite>&);
 void snap_piece_to_square(const sf::Vector2f&, PieceSprite*);
-void debug_print_vector(const sf::Vector2i&);
+void debug_print_vector(const sf::Vector2f&);
